GIS and Green Building Certification data of ongoing plans in Uppsala Sweden
The GIS data herein is derived from either public documents produced by the Uppsala Kommune (Uppsala.se) or in the case of the number of Green Certified buildings, open access info produced by the Swedish Green Building Council (SCB.se). This is information is of public interest and will be used to test the eCitizenScience, eGov, and PeoplesEPA  tools.